pyproject.toml
[project]
name = "fitbit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.12.0",
    "anthropic>=0.7.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.47.1",
]
